refactor(django_sage_qrcode): Seprate admin and serilizers and views base on qrcode model kind
