+++ django_sage_qrcode/api/serializer/barcode.py
@@ -1,0 +1,21 @@
+from rest_framework import serializers
+
+from django_sage_qrcode.models import BarcodeText, Barcode, BarcodeUrl
+
+
+class BarcodeSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Barcode
+        fields = "__all__"
+
+
+class BarcodeTextSerializer(BarcodeSerializer):
+    class Meta:
+        model = BarcodeText
+        fields = "__all__"
+
+
+class BarcodeUrlSerializer(BarcodeSerializer):
+    class Meta:
+        model = BarcodeUrl
+        fields = "__all__"
