+++ django_sage_qrcode/api/serializer/base.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from django_sage_qrcode.models import QRCode
+
+
+class QRCodeSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = QRCode
+        fields = "__all__"
