+++ django_sage_qrcode/utils/admin.py
